Refactor SlidingBanner into generic + CutoffBanner, fix animation and content offset behind fixed banner
