+++ app/page.tsx
@@ -112,8 +112,9 @@ function HomeContent() {
       <div className="min-h-screen">
         {/* Hero Section */}
         <section
-          className="relative flex flex-col overflow-hidden h-[100dvh]"
+          className="relative flex flex-col overflow-hidden"
           style={{
+            height: 'calc(100dvh - var(--banner-h, 0px))',
             background:
               'linear-gradient(165deg, #6B1D3A 0%, #4A1838 20%, #1E2D45 45%, #023047 70%, #011a28 100%)',
           }}
